mock

+++ 03_TDD/chap1_introduction/TP/Corrections/Challenge_01_Mock/src/Storable.php
@@ -10,5 +10,5 @@ interface Storable{
     // function total():float;
 
     function getStorage():array;
-    function restoreQuantity(string $name, int $quantity):void;
+    function restoreQuantity(string $name, float $quantity):void;
 }
